+++ src/camera.cpp
@@ -3,7 +3,7 @@
 Camera::Camera(Vec3 &eyePos, int imageHeight, int imageWidth) : eyePos(eyePos), imageHeight(imageHeight), imageWidth(imageWidth) {
                                                                 };
 
-Ray Camera::getRay(int i, int j)
+Ray Camera::getRay(double i, double j)
 {
 
     return Ray(eyePos, Vec3(j, 0, i) - eyePos);
